business/business-antenna-connector/src/main/java/eu/estcube/gsconnector/commands/radio/SetVFO.java
package eu.estcube.gsconnector.commands.radio;

import eu.estcube.domain.TelemetryCommand;
import eu.estcube.gsconnector.commands.HamlibCommandBuilder;

public class SetVFO implements HamlibCommandBuilder {

    public StringBuilder createMessageString(TelemetryCommand command) {
        StringBuilder messageString = new StringBuilder();
        messageString.append("+V ");
        messageString.append(command.getParameter("VFO"));
        messageString.append("\n");
        return messageString;
    }
}
